requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
WTForms
pika
cryptography
python-dotenv
requests 